+++ src/main/java/me/villagerunknown/villagercoin/Villagercoin.java
@@ -3,13 +3,13 @@ package me.villagerunknown.villagercoin;
 import me.villagerunknown.platform.Platform;
 import me.villagerunknown.platform.PlatformMod;
 import me.villagerunknown.platform.manager.featureManager;
+import me.villagerunknown.platform.util.RegistryUtil;
 import me.villagerunknown.villagercoin.component.*;
 import me.villagerunknown.villagercoin.feature.*;
 import me.villagerunknown.villagercoin.item.CoinItems;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
 import net.minecraft.block.Block;
-import net.minecraft.component.ComponentType;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemGroup;
 import net.minecraft.item.ItemStack;
@@ -23,17 +23,18 @@ import net.minecraft.util.Identifier;
 import org.slf4j.Logger;
 
 import java.util.Comparator;
-import java.util.function.UnaryOperator;
 
 public class Villagercoin implements ModInitializer {
 	
-	public static PlatformMod<VillagercoinConfigData> MOD = Platform.register( "villagercoin", Villagercoin.class, VillagercoinConfigData.class );
-	public static String MOD_ID = MOD.getModId();
-	public static Logger LOGGER = MOD.getLogger();
-	public static VillagercoinConfigData CONFIG = MOD.getConfig();
+	public static final PlatformMod<VillagercoinConfigData> MOD = Platform.register( "villagercoin", Villagercoin.class, VillagercoinConfigData.class );
+	public static final String MOD_ID = MOD.getModId();
+	public static final Logger LOGGER = MOD.getLogger();
+	public static final VillagercoinConfigData CONFIG = MOD.getConfig();
 	
 	public static final int MAX_STACK_COUNT_CAP = 1073741822;
-	public static int MAX_STACK_COUNT = 5000;
+	public static int MAX_STACK_COUNT = CONFIG.maximumCoinStackSize;
+	
+	private static boolean loaded = false;
 	
 	public static final RegistryKey<ItemGroup> ITEM_GROUP_KEY = RegistryKey.of(Registries.ITEM_GROUP.getKey(), Identifier.of(MOD_ID, "item_group"));
 	
@@ -44,8 +45,6 @@ public class Villagercoin implements ModInitializer {
 	
 	@Override
 	public void onInitialize() {
-		MAX_STACK_COUNT = CONFIG.maximumCoinStackSize;
-		
 		if( CONFIG.maximumCoinStackSize > MAX_STACK_COUNT_CAP ) {
 			Villagercoin.LOGGER.warn( "Maximum Coin Stack Size exceeds limit of " + MAX_STACK_COUNT_CAP );
 			Villagercoin.LOGGER.info( "Maximum Coin Stack Size has been set to: " + MAX_STACK_COUNT_CAP );
@@ -53,27 +52,62 @@ public class Villagercoin implements ModInitializer {
 			MAX_STACK_COUNT = MAX_STACK_COUNT_CAP;
 		} // if
 		
-		// # Initialize Mod
+		// Initialize features
 		init();
 	}
 	
 	private static void init() {
+		if( loaded ) {
+			return;
+		} // if
+		
+		// # Initialize mod with Platform
 		Platform.init_mod( MOD );
 		
 		// # Load Components
 		new Components();
 		
-		// # Activate Features
-		featureManager.addFeature( "coinCrafting", CoinCraftingFeature::execute );
-		featureManager.addFeature( "coinStackCrafting", CoinStackCraftingFeature::execute );
-		featureManager.addFeature( "receiptCrafting", ReceiptCraftingFeature::execute );
-		featureManager.addFeature( "ledgerCrafting", LedgerCraftingFeature::execute );
+		// # Activate Primary Features
+		featureManager.addFeature( "villagercoin-item-group", CustomItemGroupFeature::execute );
 		
 		featureManager.addFeature( "coin", CoinFeature::execute );
+		featureManager.addFeature( "coin-crafting", CoinCraftingFeature::execute );
 		
-		featureManager.addFeature( "structuresIncludeCoins", StructuresIncludeCoinsFeature::execute );
-		featureManager.addFeature( "mobsDropCoins", MobsDropCoinsFeature::execute );
-		featureManager.addFeature( "merchantCoinTrading", MerchantCoinTradingFeature::execute );
+		featureManager.addFeature( "collectable-coin", CollectableCoinFeature::execute );
+		featureManager.addFeature( "edible-coin", EdibleCoinFeature::execute );
+		featureManager.addFeature( "inventory-effect-coin", InventoryEffectCoinFeature::execute );
+		
+		featureManager.addFeature( "receipt", ReceiptFeature::execute );
+		featureManager.addFeature( "ledger", LedgerFeature::execute );
+		featureManager.addFeature( "receipt-crafting", ReceiptCraftingFeature::execute );
+		featureManager.addFeature( "ledger-crafting", LedgerCraftingFeature::execute );
+		
+		featureManager.addFeature( "coin-bank", CoinBankBlocksFeature::execute );
+		featureManager.addFeature( "coin-stack", CoinStackBlocksFeature::execute );
+		featureManager.addFeature( "coin-bank-crafting", CoinBankCraftingFeature::execute );
+		featureManager.addFeature( "coin-stack-crafting", CoinStackCraftingFeature::execute );
+		
+		// # Activate Supporting Features
+		featureManager.addFeature( "structures-include-coins", StructuresIncludeCoinsFeature::execute );
+		featureManager.addFeature( "mobs-drop-coins", MobsDropCoinsFeature::execute );
+		featureManager.addFeature( "merchant-coin-trading", MerchantCoinTradingFeature::execute );
+		
+		// # Activate Block Entity Loaders
+		featureManager.addFeatureLast( "coin-bank-block-entities", CoinBankBlockEntityFeature::execute );
+		featureManager.addFeatureLast( "coin-stack-block-entities", CoinStackBlockEntityFeature::execute );
+		
+		// # Load Features
+		featureManager.loadFeatures();
+		
+		loaded = true;
+	}
+	
+	public static void load() {
+		if( loaded ) {
+			return;
+		} // if
+		
+		init();
 	}
 	
 	public static TagKey<Item> getItemTagKey(String id ) {
@@ -82,6 +116,10 @@ public class Villagercoin implements ModInitializer {
 	
 	public static TagKey<Block> getBlockTagKey(String id ) {
 		return TagKey.of( RegistryKeys.BLOCK, Identifier.of(MOD_ID, id) );
+	}
+	
+	public static void addItemToGroup( Item item ) {
+		RegistryUtil.addItemToGroup( ITEM_GROUP_KEY, item );
 	}
 	
 	public static Comparator<Integer> reverseSort = new Comparator<Integer>() {