+++ src/main/java/me/villagerunknown/villagercoin/VillagercoinConfigData.java
@@ -30,6 +30,9 @@ public class VillagercoinConfigData implements me.shedaniel.autoconfig.ConfigDat
 	public boolean addCoinsToStructureLootTables = true;
 	
 	@ConfigEntry.Category("LootTables")
+	public boolean addCoinsToModdedStructureLootTables = true;
+	
+	@ConfigEntry.Category("LootTables")
 	public int copperLootTableRolls = 10;
 	
 	@ConfigEntry.Category("LootTables")
@@ -65,6 +68,9 @@ public class VillagercoinConfigData implements me.shedaniel.autoconfig.ConfigDat
 	
 	@ConfigEntry.Category("MobDrops")
 	public boolean addCoinsToMobDrops = true;
+	
+	@ConfigEntry.Category("MobDrops")
+	public boolean addCoinsToModdedMobDrops = true;
 	
 	@ConfigEntry.Category("MobDrops")
 	public boolean enableBreedableMobDrops = false;
@@ -160,6 +166,13 @@ public class VillagercoinConfigData implements me.shedaniel.autoconfig.ConfigDat
 	
 	@ConfigEntry.Category("Receipts")
 	public String defaultReceiptThankYouMessage  = "Thank you for shopping with us!";
+	
+	/**
+	 * Coin Banks
+	 */
+	
+	@ConfigEntry.Category("CoinBanks")
+	public int maximumCoinBankCurrencyValue = Integer.MAX_VALUE;
 	
 	/**
 	 * Coin Stacks