+++ src/main/java/me/villagerunknown/villagercoin/feature/CoinStackBlocksFeature.java
@@ -1,17 +1,13 @@
 package me.villagerunknown.villagercoin.feature;
 
+import me.villagerunknown.platform.list.BlocksList;
 import me.villagerunknown.platform.util.RegistryUtil;
 import me.villagerunknown.villagercoin.Villagercoin;
-import me.villagerunknown.villagercoin.block.entity.AbstractCurrencyValueBlockEntity;
 import me.villagerunknown.villagercoin.component.CurrencyComponent;
 import me.villagerunknown.villagercoin.type.CoinType;
 import net.minecraft.block.Block;
-import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.item.BlockItem;
 import net.minecraft.item.Item;
-import net.minecraft.registry.Registries;
-import net.minecraft.registry.Registry;
-import net.minecraft.util.Identifier;
 
 import static me.villagerunknown.villagercoin.component.Components.CURRENCY_COMPONENT;
 
@@ -31,47 +27,48 @@ public class CoinStackBlocksFeature {
 	public static final long EMERALD_VALUE = CoinFeature.EMERALD_VALUE;
 	public static final long NETHERITE_VALUE = CoinFeature.NETHERITE_VALUE;
 	
+	public static BlocksList blocks = new BlocksList();
+	
 	public static void execute(){}
 	
-	public static BlockEntityType<? extends AbstractCurrencyValueBlockEntity> registerCoinStackBlockEntities(BlockEntityType.Builder<? extends AbstractCurrencyValueBlockEntity> builder) {
-		// # Register the Block Entity Types
-		return Registry.register(
-				Registries.BLOCK_ENTITY_TYPE,
-				Identifier.of(Villagercoin.MOD_ID, COIN_STACK_STRING),
-				builder.build()
-		);
+	public static void addBlock( Block block ) {
+		blocks.addBlock( block );
 	}
 	
-	public static Block registerCoinStackBlock( String id, Block block, long value ) {
-		Block registeredBlock = RegistryUtil.registerBlock( id, block, Villagercoin.MOD_ID );
+	public static Block registerCoinStackBlock( String namespace, String id, Block block, long value ) {
+		Block registeredBlock = RegistryUtil.registerBlock( id, block, namespace );
 		
-		Item item = RegistryUtil.registerItem(id, new BlockItem(registeredBlock, new Item.Settings().component( CURRENCY_COMPONENT, new CurrencyComponent( value ))), Villagercoin.MOD_ID);
+		Item item = RegistryUtil.registerItem(id, new BlockItem(registeredBlock, new Item.Settings().component( CURRENCY_COMPONENT, new CurrencyComponent( value ))), namespace);
 		
-		RegistryUtil.addItemToGroup( Villagercoin.ITEM_GROUP_KEY, item );
+		Villagercoin.addItemToGroup( item );
+		
+		addBlock( registeredBlock );
 		
 		return block;
 	}
 	
-	public static Block registerFireproofCoinStackBlock( String id, Block block, long value ) {
-		Block registeredBlock = RegistryUtil.registerBlock( id, block, Villagercoin.MOD_ID );
+	public static Block registerFireproofCoinStackBlock( String namespace, String id, Block block, long value ) {
+		Block registeredBlock = RegistryUtil.registerBlock( id, block, namespace );
 		
-		Item item = RegistryUtil.registerItem(id, new BlockItem(registeredBlock, new Item.Settings().fireproof().component( CURRENCY_COMPONENT, new CurrencyComponent( value ))), Villagercoin.MOD_ID);
+		Item item = RegistryUtil.registerItem(id, new BlockItem(registeredBlock, new Item.Settings().fireproof().component( CURRENCY_COMPONENT, new CurrencyComponent( value ))), namespace);
 		
-		RegistryUtil.addItemToGroup( Villagercoin.ITEM_GROUP_KEY, item );
+		Villagercoin.addItemToGroup( item );
+		
+		addBlock( registeredBlock );
 		
 		return block;
 	}
 	
-	public static Block registerCraftableCoinStackBlock( CoinType type, String id, Block block, long value ) {
-		Block registeredBlock = registerCoinStackBlock( id, block, value );
+	public static Block registerCraftableCoinStackBlock( CoinType type, String namespace, String id, Block block, long value ) {
+		Block registeredBlock = registerCoinStackBlock( namespace, id, block, value );
 		
 		CoinStackCraftingFeature.registerCraftingResultCoinStack( type, registeredBlock, value );
 		
 		return block;
 	}
 	
-	public static Block registerCraftableFireproofCoinStackBlock( CoinType type, String id, Block block, long value ) {
-		Block registeredBlock = registerFireproofCoinStackBlock( id, block, value );
+	public static Block registerCraftableFireproofCoinStackBlock( CoinType type, String namespace, String id, Block block, long value ) {
+		Block registeredBlock = registerFireproofCoinStackBlock( namespace, id, block, value );
 		
 		CoinStackCraftingFeature.registerCraftingResultCoinStack( type, registeredBlock, value );